extract interfaces to schema file
