+++ tsconfig.json
@@ -5,6 +5,7 @@
     "sourceMap": true
   },
 "files": [
+    "src/core/schema.ts",
     "src/core/maker.ts",
     "src/core/angle.ts",
     "src/core/point.ts",
